requirements.txt
requests>=2.32.0
python-dotenv>=1.0.1
Appium-Python-Client>=4.1.0
